sampling mega is common, ability check doesnt trigger on predicted mega
