+++ fp/data/sets/smogon.py
@@ -286,9 +286,9 @@ class SmogonSets(PokemonSets):
     def move_usage_rates(self, pkmn: Pokemon) -> list[tuple[str, float]]:
         # (move, usage_rate) pairs sorted by usage; smogon stats carry no
         # move associations so this is the only moveset information available
-        return self.get_raw_pkmn_sets_from_pkmn_name(pkmn.name, pkmn.base_name).get(
-            MOVES_STRING, []
-        )
+        return self.get_raw_pkmn_sets_from_pkmn_name(
+            pkmn.name, pkmn.base_name, pkmn.mega_name
+        ).get(MOVES_STRING, [])
 
     def get_raw_count(self, pkmn_name) -> int | None:
         if pkmn_name not in self.all_pkmn_counts:
